Create main app logic
Inherited from Django AbstractUser and Group models, created serializers, additional endpoints
for adding a user to a group and removing a user from a group, user and group viewsets. Overrode
group viewset destroy method in order to be able to delete only emty groups as requested.
Hid the secret key.